cambios en copy
